Prepping by refactoring launch button out of menu
